Can now read and display a DB's contents in the application.
Not super satisifed with it since I believe it does not use a persistent db connection
but that is something I will look into in the future.

For now it works which is very nice.

FossilOrigin-Name: 989229fea71d016558f8d498c0ae312bc4b3f919ebb6001b2325f2c766ca4465