Made Database.Tagger.Access.deleteDescriptor return IO () instead of MaybeT IO ()
Not sure why but I think that fixed the issue. MaybeT seemed to be messing with the
sequencing of the additional delete DELETE executions.

At least I haven't been able to easily reproduce the problem.

I still don't know why it was happening in the first place.

Generally, the issue was:
In some circumstances sometime surrounding the deletion of a descriptor,
A new relation would be created for that descriptor for some seemingly random descriptor

I say seemingly random in that there was a pattern but I didn't want to use the word
"unrelated"

This relationship would persist after descriptor deletion so that it was a non-cascaded
reference in MetaDescriptor.

It would also not be deleted by the DELETE execution in deleteDescriptor.

So I don't now why the relation was created in the first place nor do I know why it wasn't
being deleted.

But at least making the sequencing more strict (I think) provides a solution to the result
of the problem, though not necessarily a solution to the root cause.

FossilOrigin-Name: b283c70e3ca2253ee5e4e0f480a1812213a49e8fef5e10b86d082f895077a117