Added some events and widgets to open or change db connections.
Automatic connection can be configured in the toml file and connections
can be changed on the fly while the program is running.

There is no way to tell if a connection is actually established until a Task is tried.
If a message is printed to stderr then the db is probably not connected.

Openening or changing a connection will close the old connection if it exists,
open the new connection and activate the foreign key pragma. So all of that previous
logic has been moved out of Main and into Handler.

FossilOrigin-Name: fb2e7f13a12916b1c980584622090e757e921bdc0f1e1867ef6e8b1b429c534b