Added a new config table "descriptor_tree" and option "main_request"
This field will be used to set the descriptor that the refresh button in the main
descriptor tree widget will call when pressed.

By default it is set to #META#.

This configuration is currently required to run the program but is unused as of this
commit but will be implemented in subsqequent commits.

FossilOrigin-Name: 15c3a213aac3f877117ad7ebb2b42ca2627c63b0134682138995b249562c8471