NOTE - This branch is to develop a representative file system.
A representative file is linked to a descriptor.

A descriptor may have 0 or 1 rep. files.
	They can be changed.

Whatever is the preview method for the files used, there must be a thumbnail-like
mode for referencing rep. files.

Rep. files are meant to be aids in tagging. When the purpose or meaning of a descriptor
may be hard to discern, the representative file is used as a reference to understand
the given descriptor better.

A file may be representative for any number of descriptors.

FossilOrigin-Name: 932939593c0cddf0aacfe3dfd850c3781f8b6639beecac1c3e9e998f79b85269