Created a RootedDescriptorTree data type
Added lenses for it, unimplemented in the models.

RootedDescriptorTree :

A data type to be used in Tagger sub-models

makes use of lenses

Contains a text field which has the name of a descriptor. This is used when refreshing
the tree to make sure the correct root descriptor is fetched from the database.

The Ord instance is only an instance of the underlying DescriptorTree and ignores the
name field entirely.

FossilOrigin-Name: b1f7b2bb195f886bc1dbc3e1a020e348aacfee6084b18ed29a90fdcade3eb576