Lazy file loading works using a BufferList
The part that doesn't work is cycling. My cPop and cDequeue methods for the BufferList
are confusing and mess with the ordering all the time.

So I will change that in a future commit to cycle on the total virtual list
and not the separate _buffer + _list which is unintuitive for a user.

But the program doesn't crash and is usable if confusing, hence the commit.

FossilOrigin-Name: d342e66c0c89115d1ec83a26082d63dbbe39b9bf898e6af91d6905993f0fb626