Implemented queries in Access to take advantage of the new view FileWithTags.
Created a newtype wrapper for DatabaseFileWithTags that indicates that a DatabaseFileWithTags
  actually has all of its tags. This type can only be returned by looking up a FileKey
  in the FileWithTags view.

Any other queries on FileWithTags only return FileKeys to reflect this.
  Because of this, many usages of FileWithTags may actually be unnecessary.

Wrote a CTE for walking an infraTree, which I plan to use to find FWT's based on Relation.
  created a type synonym QueryRequiring a = Query for concattable queries that provides
  a usage hint, for the programmer, on what the additional query requires as far as parameter substitution.

Created a DescriptorPtr newtype which is a just a wrapper for DescriptorKey.

Adjusted the reduce function in RowMap to operate on (Map File (IntSet TagKey))
  to avoid duplication of tags in the resulting FWT's. Adjusted the type synonym accordingly.

FossilOrigin-Name: e001a5b5a003e6f1e775b11e214232e40c7601696488db18da7446c8147749e5