Changed TaggedConnection._taggedconnectionConnInstance from (Maybe BareConnection) to BareConnection.
This simplifies some things quite a bit and there is not really a reason for it to be
Maybe anyways. If the connection fails for any reason then that error is already reported
by sqlite. There's no reason to introduce a redundant layer (except for retrying a connection
I suppose but that's not very important to me right now.)

FossilOrigin-Name: 35f39be53e70fae9fa86ff2761014899973b4fd0e9af92bd97d5b503283154be