Lots of changes while writing parsers: Changed AST structs and wrote some more parsers
Gave TermTree a new constructor Bottom, which is the same as Simple.
This constructor is used for pattern matching the last term in a relation query.

Added a boolean argument to the term tree parsers to differentiate when a term is 
Simple vs Bottom.

Changed Clause to CombinableTerm and made it hold a single TermTree instead of a list

Changed Sentence to just hold a list of CombinableTerms

Added a new data type CombinableSentence which is like CombinableTerm but for Sentences.

Updated/wrote queries for all structs above.

FossilOrigin-Name: e5aaace257d0bb44da7d28d68284ec431a8327c07de09a9dd98c8c595471569c