Made the event handler block events for ToggleDescriptorLeafVisibility if the given descriptor is a hash descriptor ie something like #ALL# that is bookended by hashes.
This makes it possible to make descriptors immutable in the UI but not possible to
rename them without either using the taggerlib or sqlite. Not sure if this is ideal
but it works for now.

I want to at least block renaming #ALL# #META# and #UNRELATED# in the lib itself
but still give users the option to make their own tags immutable as well.

FossilOrigin-Name: ea63863b67171f7548ea18468213e1d359501e147de3a0fe6fe71c72775baeac