Modifying the Visibility data structure to be a little more useful.
Gave it a new constructor for widgets with panes, this constructor takes a main
visibility and a set of visibilities to specify when and what panes should be visible.
Probably not the best way to do it, I don't think there needs to be this many constructors
for visibility but it's alright, just gives you something more to work with.

FossilOrigin-Name: 1d78787f26cde4e769ffa83cd99cbd31427b0ad1f39568289a431bd20dac63bb