Changed the default SetOp in the combinableSentenceParser from Union to Intersect
This change required a lot of thought though I think it should have been obvious
from the start.

Given a query: a d| (b u| c) d e

The current implementation is confusing and ambiguous. This query actually consists of 
three parts: "a", "b u| c", and "d e".
Given that the current default SetOp is a Union and sentence combinations are left-associative
this query could be conecptualized as: (a - (b + c)) + (d // e) where (//) is an interesection.

The reason that this is confusing is because it requires a query-writer to understand
how SentenceTrees are created and grouped by the parser. They must know that
that the ending of a SentenceBranch followed by no explicit SetOp would create a SentenceNode
from any terms that follow. They must also know that all terms in a SentenceNode's SentenceSet's Sentence
are interected by default. And they must also know that, during final combination in the query engine,
combinations are left-associative and the default combination SetOp is a Union. So it
becomes very important to just know that the third portion of the query, "d e", is
actually a distinct grouped query by itself like: "u|(d e)" and this is not
immediately obvious from the way it was entered.

By changing this default SetOp to Intersect, this structural distinction is not lost,
but does not become information vital to the creation of an accurate query.
Though the last part of the query is still technically "i|(d e)", because of the
fact that any given Sentence's terms are intersected by default, the final part of the 
query can actually just be thought of as two query parts: "d" and "e". This also allows
us to logically comprehend any query by just substituting branches (parenthesized portions)
with names and outer parentheses to better understand grouping.

So that the query can become: (a d| t') d e, where t': b u| c

and since the query now contains another branch, we can do the same again: t'' d e, where
t'': a d| t'

From these substitutions, it becomes more clear that query is really just a series of
intersections of terms.

As an aside, if we were to perform this same substitution with the current implementation
we would end up with: t'' u| q', where q': d // e

but given that the original query had implicit grouping and not an explicit parenthesized branch
for the last portion, it is visually ambiguous on if this could actually be: (q' u| d) // e

tl;dr
changing the default SetOp in the parser makes the query language more logically consistent.

FossilOrigin-Name: 3bfb845ac2b853a7ff49569a4a68cbbe5a32fde553fb8faa4c6d718559250bd4