Fixed the trailing space issues by using a modified version of the chainl1 parser.
This version applies a default function to the lhs operand in the event that
a rhs operand fails to parse. Therefore, the entire operator application
won't fail if it gets all the way to the end of an expression and finds a trailing space,
which signals an implicit setop, but no rhs. It then just returns the lhs and ignores 
the rest.

FossilOrigin-Name: e6589a357e09700d12b489249442bf68fa0be26996908bdc38aca57c62e0399b