Changed traverseHierarchy to be more uniform in how it applies its transformation function.
The initial application of the function is when all parent nodes are gathered.
Previously, the subsequent applications where only on when a given node had any number of children.

If the transformation function is some sort of filter, then this could lead to unintended behavior.

So now the transformation is applied immediately after finding the children for each given node.
This way, if it is a filter, then children are filtered before a node is determined to be
a leaf or branch. Which I believe is what the expected behavior is.

FossilOrigin-Name: 0281c92b0569bbb38f7a369d0ce02fd4e9cd11632731778850c9207e3626d7aa