Moved the query editor widget to an entirely new pane which is laid over the entire window.
This pane is active when an "edit" button is pressed and can be exited. Depending on
whether or not it is active will affect the behavior of certain portions of the UI.

I moved it to have more room to play around with widgets instead of just
restricting the query builder widgets to the upper left-hand corner.

FossilOrigin-Name: 6c24d794f71f8ab82a5165395590b6d04ed018f33f062f9dbc6a7abb4b94aa2b