Fixed the existing query engine test cases.
The issue was that a TagExpression was being simplified to a ring, resolved to
an expression of filesets, then the ring was evaluated.

To fix the tests, I changed it so that a single (TagExpression (DTerm Pattern)) is
always evaluated to (HashSet Tag) before being resolved to a (HashSet File). This means
that any ring expression inside of a TagExpression is always going to correspond to tags.
So the constructure (RingExpression (MagmaExpression a)) can probably be reduced to
just (a).

FossilOrigin-Name: 57fc3287bffb9af64c88b65ba1e83b07659f974f3a4763a5f586c3fc06e43bb9