Removing code associated with the old expressions. Replacing it in positions that are easy to replace.
Most of the code for the query builder and its associated events have to be rebuilt
before this branch can be closed. Refer back to this check-in to see all places that
need to be rebuilt.

FossilOrigin-Name: 022207ffe06dcc6a06c103d5fa6f415efafaa7f30dcb60c0ea528cb809bdab5c