Fixed the tests by distributing the generated TagExpressions.
The Arbitrary instance can potentially create a tag expression that is a single term
but wrapped in an arbitrary number of identity constructors like: Magma, Ring, TagValue,
and so on. 

Because this introduces structural inequality, but not computational inequality.
Simply distributing the expressions, which is a form of normalization, allows
the two to be compared.

FossilOrigin-Name: 2c8722e801db0c92a11828ed2e6aa0ebf7dd40c7e308aa07429f2725f0b1e56f